Split min_quad_with_fixed implementation into separate cpp.
